requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
cloudpickle==1.6.0
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
gym==0.19.0
h5py==3.1.0
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.6.0.66
openTSNE==0.6.0
opt-einsum==3.3.0
pandas==1.1.5
patsy==0.5.6
pgmpy==0.1.24
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
stable-baselines3==1.3.0
statsmodels==0.12.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.18
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
